Restructured Sentry integration tabs and components
