+++ .changeset/modern-parents-work.md
@@ -1,0 +1,5 @@
+---
+'@spotlightjs/overlay': minor
+---
+
+Restructured Sentry integration tabs and components
