+++ packages/overlay/src/integrations/sentry/components/insights/QuerySummary.tsx
@@ -112,15 +112,12 @@ const QuerySummary = ({ showAll }: { showAll: boolean }) => {
           {filteredDBSpans.map(span => (
             <tr key={span.span_id} className="hover:bg-primary-900">
               <td className="text-primary-200 w-2/5 truncate whitespace-nowrap px-6 py-4 text-left text-sm font-medium">
-                <Link className="truncate hover:underline" to={`/explore/traces/${span.trace_id}`}>
+                <Link className="truncate hover:underline" to={`/traces/${span.trace_id}`}>
                   {truncateId(span.trace_id)}
                 </Link>
               </td>
               <td className="text-primary-200 w-[15%] whitespace-nowrap px-6 py-4 text-right text-sm font-medium">
-                <Link
-                  className="truncate hover:underline"
-                  to={`/explore/traces/${span.trace_id}/spans/${span.span_id}`}
-                >
+                <Link className="truncate hover:underline" to={`/traces/${span.trace_id}/spans/${span.span_id}`}>
                   {truncateId(span.span_id)}
                 </Link>
               </td>
