+++ packages/overlay/src/integrations/sentry/components/insights/envelopes/EnvelopeList.tsx
@@ -11,8 +11,8 @@ import { useSentryEnvelopes } from '../../../data/useSentryEnvelopes';
 import { useSentryHelpers } from '../../../data/useSentryHelpers';
 import { sdkToPlatform } from '../../../utils/sdkToPlatform';
 import { truncateId } from '../../../utils/text';
-import HiddenItemsButton from '../../HiddenItemsButton';
-import PlatformIcon from '../../PlatformIcon';
+import HiddenItemsButton from '../../shared/HiddenItemsButton';
+import PlatformIcon from '../../shared/PlatformIcon';
 import EnvelopeDetails from './EnvelopeDetails';
 
 export default function EnvelopeList() {
@@ -50,7 +50,7 @@ export default function EnvelopeList() {
                 return null;
               }
               return (
-                <Link key={envelopeEventId} to={`/explore/envelopes/${header.event_id}`}>
+                <Link key={envelopeEventId} to={`/insights/envelopes/${header.event_id}`}>
                   <div
                     className={classNames(
                       'hover:bg-primary-900 border-b-primary-900 flex cursor-pointer items-center gap-4 border-b px-6 py-2 transition-all',
