+++ packages/overlay/src/integrations/sentry/components/insights/envelopes/index.tsx
