+++ packages/overlay/src/integrations/sentry/components/insights/sdks/index.tsx
