+++ packages/overlay/src/integrations/sentry/components/shared/HiddenItemsButton.tsx
