+++ packages/overlay/src/integrations/sentry/components/shared/PlatformIcon.tsx
