+++ packages/overlay/src/integrations/sentry/components/shared/Resizer.tsx
