+++ packages/overlay/src/integrations/sentry/components/shared/SpanResizer.tsx
