+++ packages/overlay/src/integrations/sentry/components/traces/TraceDetails/components/TraceDetailHeader.tsx
@@ -1,4 +1,4 @@
-import type { Trace } from '../../../../../types';
+import type { Trace } from '../../../../types';
 import TraceIcon from '../../TraceIcon';
 import { TraceRootTxnName } from './TraceRootTxnName';
