+++ packages/overlay/src/integrations/sentry/components/traces/spans/SpanDetails.tsx
@@ -2,15 +2,15 @@ import { type ReactNode, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { format as formatSQL } from 'sql-formatter';
 import Table from '~/ui/Table';
-import JsonViewer from '../../../../../../components/JsonViewer';
-import SidePanel, { SidePanelHeader } from '../../../../../../ui/SidePanel';
-import { DB_SPAN_REGEX } from '../../../../constants';
-import dataCache, { isErrorEvent } from '../../../../data/sentryDataCache';
-import type { SentryErrorEvent, Span, TraceContext } from '../../../../types';
-import { formatBytes } from '../../../../utils/bytes';
-import { getFormattedDuration } from '../../../../utils/duration';
-import { ErrorTitle } from '../../../events/error/Error';
-import DateTime from '../../../shared/DateTime';
+import JsonViewer from '../../../../../components/JsonViewer';
+import SidePanel, { SidePanelHeader } from '../../../../../ui/SidePanel';
+import { DB_SPAN_REGEX } from '../../../constants';
+import dataCache, { isErrorEvent } from '../../../data/sentryDataCache';
+import type { SentryErrorEvent, Span, TraceContext } from '../../../types';
+import { formatBytes } from '../../../utils/bytes';
+import { getFormattedDuration } from '../../../utils/duration';
+import { ErrorTitle } from '../../events/error/Error';
+import DateTime from '../../shared/DateTime';
 import SpanTree from './SpanTree';
 
 function DBSpanDescription({ desc, dbType }: { desc: string; dbType?: string }) {
