+++ packages/overlay/src/integrations/sentry/components/traces/spans/SpanTree.tsx
@@ -1,5 +1,5 @@
-import classNames from '../../../../../../lib/classNames';
-import type { Span, TraceContext } from '../../../../types';
+import classNames from '../../../../../lib/classNames';
+import type { Span, TraceContext } from '../../../types';
 import SpanItem from './SpanItem';
 
 export default function SpanTree({
