--- packages/overlay/src/integrations/sentry/tabs/PerformanceTab.tsx
@@ -1,14 +1,0 @@
-import PerformanceTabDetails from '../components/performance';
-import { SentryEventsContextProvider } from '../data/sentryEventsContext';
-
-import { Route, Routes } from 'react-router-dom';
-
-export default function PerformanceTab() {
-  return (
-    <SentryEventsContextProvider>
-      <Routes>
-        <Route path="/*" element={<PerformanceTabDetails />} />
-      </Routes>
-    </SentryEventsContextProvider>
-  );
-}
